requirements.txt
snowflake-connector-python[pandas]>=3.7,<4
pandas>=2.0,<3
plotly>=5.18,<6
